fix: dedupe zone label helper and improve tamil no-results
